requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1

SQLAlchemy==2.0.32
asyncpg==0.29.0

google-genai
pillow
